relocate jd access code
