+++ refiner/app/db/jurisdictions/model.py
@@ -1,0 +1,11 @@
+from pydantic import BaseModel
+
+
+class DbJurisdiction(BaseModel):
+    """
+    Jurisdiction info.
+    """
+
+    id: str
+    name: str
+    state_code: str
